builder uses same appox function
